Warning Page Added
WarningPage.tsx added and index.tsx returned to home page.
